+++ core-backend/wasmi/src/memory.rs
@@ -30,11 +30,11 @@ use wasmi::{core::memory_units::Pages, Memory as WasmiMemory, Store, StoreContex
 
 pub fn read_memory_as<D: Decode + MaxEncodedLen>(
     memory: &impl Memory,
-    ptr: usize,
+    ptr: i32,
 ) -> Result<D, MemoryError> {
     let mut buffer = vec![0u8; D::max_encoded_len()];
     memory
-        .read(ptr, &mut buffer)
+        .read(ptr as usize, &mut buffer)
         .map_err(|_| MemoryError::OutOfBounds)?;
     let decoded = D::decode_all(&mut &buffer[..]).map_err(|_| MemoryError::MemoryAccessError)?;
     Ok(decoded)
